`src/gen_class_innards_defines.hpp`: support `constexpr` versions `containers::CxPrevCurrPair`: add class `containers::CxMoveOnlyPrevCurrPair`: add class fix various things misc. other changes

+++ src/misc/misc_defines.hpp
@@ -1,14 +1,15 @@
 #ifndef liborangepower_misc_defines_hpp
 #define liborangepower_misc_defines_hpp
-
-
+//--------
+#include <type_traits>
+//--------
 #define WIDTH2MP(some_width) ((some_width) - 1)
 #define BPRANGE2MASK(bit_pos_hi, bit_pos_lo) \
-	(((size_t)1 << (((size_t)bit_pos_hi) - (size_t)(bit_pos_lo) \
-	+ (size_t)1)) - (size_t)1)
+	(((uintmax_t)1 << (((uintmax_t)bit_pos_hi) - (uintmax_t)(bit_pos_lo) \
+	+ (uintmax_t)1)) - (uintmax_t)1)
 #define BPRANGE2SHIFTED_MASK(bit_pos_hi, bit_pos_lo) \
-	((((size_t)1 << ((size_t)(bit_pos_hi) - (size_t)(bit_pos_lo) \
-	+ (size_t)1)) - (size_t)1) << (size_t)(bit_pos_lo))
+	((((uintmax_t)1 << ((uintmax_t)(bit_pos_hi) - (uintmax_t)(bit_pos_lo) \
+	+ (uintmax_t)1)) - (uintmax_t)1) << (uintmax_t)(bit_pos_lo))
 
 
 #define _ALIGN4 aligned(4)
@@ -28,6 +29,5 @@
 
 #define TOKENPASTE3(x, y, z) TOKENPASTE2(x, TOKENPASTE2(y, z))
 #define TOKENPASTE4(x, y, z, w) TOKENPASTE3(x, y, TOKENPASTE2(z, w))
-
-
+//--------
 #endif		// liborangepower_misc_defines_hpp
